Initial commit
VBIT2 is the vbit-pi-stream code rewritten in C++. It should be able to
cope with carousels and won't need to rely on a restart when pages are
added or removed.  Carousels will also have better control of timing.
This is just a copy for safe keeping. It does nothing useful yet.