Implemented Delete
Delete works like this. FileMonitor scans for new, changed or deleted
files. If it sees any changes it updates the page list. Delete is
especially complicated as we want to avoid threads from crashing each
other, by deleting an object that the Service thread is still using.
Ideally needs a mutex but with careful timing it should work OK. At
least most of the time.